% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeData.R
\name{makeData}
\alias{makeData}
\title{Make basic wind speed data for 5 largest cities in Norway}
\usage{
makeData(seed = 10)
}
\arguments{
\item{seed}{number}
}
\value{
data frame
}
\description{
Make basic wind speed data for 5 largest cities in Norway
}
\examples{
df <- makeData()
}
