% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictSpeed.R
\name{mphtokmph}
\alias{mphtokmph}
\title{Changes a wind speed from mph to kmph}
\usage{
mphtokmph(windspeed)
}
\arguments{
\item{windspeed}{number}
}
\value{
number
}
\description{
Changes a wind speed from mph to kmph
}
\examples{
mphtokmph(10)
}
