% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictSpeed.R
\name{plotWindSpeed}
\alias{plotWindSpeed}
\title{Check data}
\usage{
plotWindSpeed(data)
}
\arguments{
\item{data}{data frame of wind speed data}
}
\value{
ggplot object
}
\description{
Check data
}
\examples{
plotWindSpeed(data)

}
